movie-rag-chatbot/requirements.txt
streamlit>=1.31.0
openai>=1.12.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
pypdf>=4.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
python-docx>=1.1.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
lxml>=5.1.0
